requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-appconf==1.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-imagekit==4.0.2
djangorestframework==3.12.2
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
pilkit==2.0
Pillow==8.1.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
